refactor(issue69): move collectors under src/data/collectors
- Move financial_collector and investor_trading_collector implementations under src/data/collectors\n- Keep src/*.py wrappers for backward-compatible imports\n- Update pipeline batch orchestrator to import implementation modules\n- Update issue69 refactor tracking todo
